COM-1675 - add contracts end on 123paie exports
